Split Lathe into a Base class and an ActiveRecord-specific class, in preparation for adding DataMapper support.
Next step will be to split up the specs along the same lines, and get rid of the requirement for ActiveSupport from the non-ActiveRecord code.

It's not clear yet what functionality should be in Lathe::Base and what functionality should be in the subclasses, but hopefully that will become apparent as a few get built.